refactor(context): ts refactor for Global, I18n, RemoteConfig context
